+++ protos/protos.js
@@ -11168,11 +11168,11 @@
     
                     /**
                      * SeekRequest snapshot.
-                     * @member {string} snapshot
+                     * @member {string|null|undefined} snapshot
                      * @memberof google.pubsub.v1.SeekRequest
                      * @instance
                      */
-                    SeekRequest.prototype.snapshot = "";
+                    SeekRequest.prototype.snapshot = null;
     
                     // OneOf field names bound to virtual getters and setters
                     var $oneOfFields;
@@ -13629,11 +13629,11 @@
     
                     /**
                      * ValidateMessageRequest name.
-                     * @member {string} name
+                     * @member {string|null|undefined} name
                      * @memberof google.pubsub.v1.ValidateMessageRequest
                      * @instance
                      */
-                    ValidateMessageRequest.prototype.name = "";
+                    ValidateMessageRequest.prototype.name = null;
     
                     /**
                      * ValidateMessageRequest schema.
@@ -14391,43 +14391,43 @@
     
                 /**
                  * HttpRule get.
-                 * @member {string} get
+                 * @member {string|null|undefined} get
                  * @memberof google.api.HttpRule
                  * @instance
                  */
-                HttpRule.prototype.get = "";
+                HttpRule.prototype.get = null;
     
                 /**
                  * HttpRule put.
-                 * @member {string} put
+                 * @member {string|null|undefined} put
                  * @memberof google.api.HttpRule
                  * @instance
                  */
-                HttpRule.prototype.put = "";
+                HttpRule.prototype.put = null;
     
                 /**
                  * HttpRule post.
-                 * @member {string} post
+                 * @member {string|null|undefined} post
                  * @memberof google.api.HttpRule
                  * @instance
                  */
-                HttpRule.prototype.post = "";
+                HttpRule.prototype.post = null;
     
                 /**
                  * HttpRule delete.
-                 * @member {string} delete
+                 * @member {string|null|undefined} delete
                  * @memberof google.api.HttpRule
                  * @instance
                  */
-                HttpRule.prototype["delete"] = "";
+                HttpRule.prototype["delete"] = null;
     
                 /**
                  * HttpRule patch.
-                 * @member {string} patch
+                 * @member {string|null|undefined} patch
                  * @memberof google.api.HttpRule
                  * @instance
                  */
-                HttpRule.prototype.patch = "";
+                HttpRule.prototype.patch = null;
     
                 /**
                  * HttpRule custom.